feat(api): implement POC database connection
