+++ apps/api/.env.test
@@ -28,4 +28,4 @@ NI_DB_MYSQL_PASSWORD = ""
 NI_DB_MYSQL_DATABASE, = ""
 NI_DB_MYSQL_HOST = ""
 NI_DB_MYSQL_PORT = ""
-NI_DB_MYSQL_DIALECT = "
+NI_DB_MYSQL_DIALECT = "mysql"
