+++ apps/api/src/server/config/schema.js
@@ -26,13 +26,5 @@ export default joi.object({
 	}),
 	cwd: joi.string(),
 	featureFlags: joi.object().pattern(/featureFlagBoas\d+[A-Za-z]+/, joi.boolean()),
-	serviceBusEnabled: joi.boolean().optional(),
-	wordpressDatabase: {
-		username: joi.string(),
-		password: joi.string(),
-		database: joi.string(),
-		host: joi.string(),
-		port: joi.string(),
-		dialect: joi.string()
-	}
+	serviceBusEnabled: joi.boolean().optional()
 });
