+++ apps/api/src/server/migration/migrator.service.js
@@ -3,17 +3,6 @@ import addFormats from 'ajv-formats';
 import { readdirSync, readFileSync } from 'node:fs';
 import { migrateNsipProjects } from './migrators/nsip-project-migrator.js';
 
-const schemas = readdirSync('./src/message-schemas/events')
-	.filter((file) => file.endsWith('.schema.json'))
-	.map((file) => {
-		console.info('adding', file);
-		return JSON.parse(readFileSync(`./src/message-schemas/events/${file}`).toString());
-	});
-
-const ajv = new Ajv({ schemas });
-
-addFormats(ajv);
-
 /**
  * @callback Migrator
  * @param {any[]} models
@@ -28,12 +17,6 @@ addFormats(ajv);
 
 const migrationMap = new Map();
 
-// TODO: Configure mappings for all models
-migrationMap.set('nsip-project', {
-	validator: ajv.getSchema('nsip-project.schema.json'),
-	migrator: migrateNsipProjects
-});
-
 /**
  *
  * @param {string} modelType
@@ -41,9 +24,31 @@ migrationMap.set('nsip-project', {
  * @returns {MigrationMap | null}
  */
 export const getMigratorForModel = (modelType) => {
+	if (migrationMap.size === 0) {
+		initializeMapping();
+	}
+
 	if (!migrationMap.has(modelType)) {
 		return null;
 	}
 
 	return migrationMap.get(modelType);
+};
+
+// the mappings are lazily initialised to improve test performance
+const initializeMapping = () => {
+	const schemas = readdirSync('./src/message-schemas/events')
+		.filter((file) => file.endsWith('.schema.json'))
+		.map((file) => {
+			return JSON.parse(readFileSync(`./src/message-schemas/events/${file}`).toString());
+		});
+
+	const ajv = new Ajv({ schemas });
+
+	addFormats(ajv);
+
+	migrationMap.set('nsip-project', {
+		validator: ajv.getSchema('nsip-project.schema.json'),
+		migrator: migrateNsipProjects
+	});
 };
