+++ apps/functions/applications-migration/local.settings.json
