+++ apps/functions/applications-migration/project-updates-migration/index.js
@@ -1,12 +1,12 @@
 import { ProjectUpdatesMigration } from './src/project-updates-migration.js';
-import { PrismaClient } from '#db-client';
 
 /**
  * @param {import('@azure/functions').Context} context
  * @param {import('@azure/functions').HttpRequest} req
  */
 export default async function (context, req) {
-	const dbClient = new PrismaClient();
-	const handler = new ProjectUpdatesMigration({ dbClient });
+	console.info(JSON.stringify(req));
+
+	const handler = new ProjectUpdatesMigration();
 	await handler.handleRequest(context, req);
 }
